pyproject.toml
[tool.poetry]
name = "custom-data-example"
version = "0.1.0"
description = ""
authors = ["Martin Stone <martindstone@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
requests = "^2.32.3"
typer = "^0.15.1"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
custom-data-example = "custom_data_example.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
